Worker scope was doing weird things. handle 0 bytes read.
