+++ src/main/java/cs455/scaling/server/Server.java
@@ -66,11 +66,11 @@ public class Server {
                     }
                     if (key.isAcceptable()) {
                         SocketChannel clientSocket = serverSocket.accept();
-                        System.out.println("New connection attempt, making a new task.");
-                        System.out.println(clientSocket);
+                        //System.out.println("New connection attempt, making a new task.");
+                        //System.out.println(clientSocket);
                         this.threadPoolManager.addTask(new REGISTER_CLIENT(this.selector, clientSocket, key, this));
                     } if (key.isReadable()) {
-                        System.out.println("New message, making a new task.");
+                        //System.out.println("New message, making a new task.");
                         this.threadPoolManager.addTask(new HANDLE_TRAFFIC(key, this));
                     }
                     iterator.remove();
@@ -153,8 +153,8 @@ public class Server {
         synchronized (clientStatistics) {
             clientStatistics.put(clientAddress, 0);
         }
-        System.out.println("Registered a client in server.");
-        System.out.println(selector.keys());
+        //System.out.println("Registered a client in server.");
+        //System.out.println(selector.keys());
     }
 
     public void deregisterClient(SocketAddress clientAddress) {
